[software] Rewrite main of Cholesky and Cholesky scheduled
