+++ software/apps/choldec_schedule/define.h
@@ -1,0 +1,43 @@
+// Copyright 2021 ETH Zurich and University of Bologna.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+
+// Author: Marco Bertuletti, ETH Zurich
+
+#define N_BANKS (NUM_CORES * 4)
+
+/* Matrix dimension */
+#define N 4
+
+//#define SINGLE
+//#define PARALLEL
+//#define LINSOLVER
+
+//#define N_COL 1024
+//#define N_ROW 1
+//#define N_DEC 2
+
+int32_t A_matrix[N * N] __attribute__((aligned(N_BANKS), section(".l1")));
+int32_t AT_matrix[N * N] __attribute__((aligned(N_BANKS), section(".l1")));
+int32_t M_matrix[N * N] __attribute__((aligned(N_BANKS), section(".l1")));
+
+#if defined(SINGLE)
+int32_t L_matrix[N * N] __attribute__((aligned(N_BANKS), section(".l1")));
+int32_t In[N] __attribute__((aligned(N_BANKS), section(".l1")));
+
+#elif defined(PARALLEL)
+int32_t LL_matrix[N_ROW * N * N_BANKS]
+    __attribute__((aligned(N_BANKS), section(".l1")));
+int32_t LR_matrix[N_ROW * N * N_BANKS]
+    __attribute__((aligned(N_BANKS), section(".l1")));
+int32_t In_matrix[N * N_BANKS]
+    __attribute__((aligned(N_BANKS), section(".l1")));
+#elif defined(LINSOLVER)
+int32_t LL_matrix[N_ROW * N * N_BANKS]
+    __attribute__((aligned(N_BANKS), section(".l1")));
+int32_t LR_matrix[N_ROW * N * N_BANKS]
+    __attribute__((aligned(N_BANKS), section(".l1")));
+int32_t In_matrix[N * N_BANKS]
+    __attribute__((aligned(N_BANKS), section(".l1")));
+int32_t In[N_BANKS] __attribute__((aligned(N_BANKS), section(".l1")));
+#endif
